+++ source/raw/bin/x64/plugins/cyber_engine_tweaks/mods/Character Preset Manager (CET)/init.lua
@@ -3,6 +3,7 @@ local runtime = assert(require("modules/runtime"),
 require("modules/constants")
 require("modules/state")
 require("modules/filesystem")
+require("modules/native_catalog")
 require("modules/catalog")
 require("modules/presets")
 require("modules/folders")